move fetchProfile to fetchData.js. myOwnProject displays profile First and Last name
